Fix some warnings

+++ src/os/async_windows.cpp
@@ -48,7 +48,7 @@ Async::Internal::WorkerResult Async::Internal::worker() {
     // point so mutually-exclusive access is guaranteed.
     if (!overlapped) return resultError();
     auto& result = toResult(overlapped);
-    result.res = numBytes;
+    result.res = static_cast<int>(numBytes);
 
     // Pass any failure back to the calling coroutine
     if (!ret) result.error = System::getLastError();
